Added random seed argument.
When `use_rand` is enabled, the fact that token indices are being
generated using the default generator results in future rand calls to be
offset. This appears to be the reason why in some cases, ToMe generates
images that are slightly different from what SD would generate with the
same random seed.

While some non-determinism is unavoidable on CUDA specifically, this can
be avoided by using an independent rand generator for that one step.

With this commit, the behavior remains unchanged if `use_rand` is enabled,
just to ensure consistency. However, if `rand_seed` is set to any
integer, the independent generator will be used and images will start to
look more like what SD would produce.